test: add PowerShell test script
The script is cross-platform but added to support Windows.
On Windows, `make test` currently fails.
It happens because `go build -o foo`
doesn't mean `go build -o foo.exe`.
This is a defensible design decision for a compiler
with strong support for cross-compilation.
All the same, it interacts poorly with makefiles.
Rather than work around it by introducing an `EXT` variable
in the makefile,
we can write a separate build plus test script.