feat(cli): switch to custom argument parser
There are issues with Kong 1.2.1 that affect our usage:

- Negative numbers are treated as flags.
- `passthrough` passes through a flag as the command.
  There is no apparent way to express "validate flags
  but pass through everything after the first positional argument".

Attempts to port recur to other CLI libraries
have surfaced issues with each.
The issues center around passthrough and a lack of customization
(for example, no way to hide a default value).
Unfortunately, it seems that no popular Go CLI library
can fully replicate what we did with argparse in Python.

I have written a custom parser to suit our needs.
Because these needs are pretty simple
(for example, there are no subcommands
and there is no mutual exclusion),
a bespoke parser should be easier to maintain
than a set of overrides for an existing parser.

With the new parser, `--foo=bar` arguments are no longer accepted.
Use `--foo bar`.

Improve the formatting of durations.

Remove the "command starts with '-'" warning.