feat(cli): semantically indent help message parts
Let's copy the indentation style of clap 4 for Rust.
It is common enough,
more readable for me than what we currently have,
and easier to imitate without automatic text indentation
than, for example, the style of Python's argparse.

Switch to single quotes.
They are a better default on the command line
because double quotes imply interpolation.